Only install to virtual environment
Remove all tasks and variables related to toggling between installation
of zaqar inside or outside of a Python virtual environment.
Installing within a venv is now the only supported deployment.

Additionally, a few changes have been made to make the creation of the
venv more resistant to interruptions during a run of the role.
* unarchiving a pre-built venv will now also occur when the venv
  directory is created, not only after being downloaded
* virtualenv-tools is run against both pre-built and non pre-built venvs
  to account for interruptions during or prior to unarchiving

Change-Id: I4a67a53d3d685c1e2f937114352569c9f45a74a0
Implements: blueprint only-install-venvs