[WIP] Initial Zuul v3 role jobs
This patch implements an initial set of jobs intended to match
the current job execution method. It does not intend to improve
how the jobs are executed - only to replicate what is currently
in openstack-infra/openstack-zuul-jobs and provide the platform
to iterate on.

[WIP] On top of that I am trying to uniformize what's in the
usual tests folder. This may fix what's happening in gates not
happening on my machine.

Change-Id: I78c32b740fc291b7373ffa0c19015946decaf743