Only replace python when re-initializing the venv
We reinitialize the venv to ensure that the right version of
python is in the venv, but we do not want virtualenv to also
replace pip, setuptools and wheel so we tell it not to. If we
don't do this then virtualenv will install the latest available
version, which is not what we want.

Change-Id: I8772e90fa3ff2653836870921ee0e182150715e4
Partial-Bug: #1764470
(cherry picked from commit da4ea69eb2de6025822c60f8678404f79126f8f4)